Rewired org remember templates.
Rewired and rewrote the org remember templates -- the previous
templates were too many and therefore too unwieldly to use, especially
since I'm moving to a one-file-per-project style.

With this rewiring, I only have to choose a particulary *category* to
remember, and it gets dumped to an inbox file. I then have to manually
move the task/note/whatever to its proper place later on, which is
what I do with tasks anyway.