Explicitly set `session-use-package`
This is inevitably set by session-initialize, but it ends up being
saved in 001-custom.init.el, as I believe it's being set via
`customize-set-variable`. When bootstrapping a new Emacs installation,
the fact that it's being set in `custom-set-variables` causes things
to fail as the `session` package wouldn't yet be installed (it's
installed when navigation.init.el is loaded); to fix that, have an
explicit `setq` while configuring the `session` package.