Add various elfeed creature comforts
This sets up the default search filter to view the last 2 weeks of
unread entries, and sets keybindings in `elfeed-search` so I can
toggle between bookmarked entries and unread entries.

I also add a binding in `elfeed-entry` to be able to quickly tag
entries for bookmarking by overloading the bookmark binding (<f7>).