Apply color fixup on solarized theme
Apparently (and I hadn't noticed for so long), I've been running the
theme slightly wrong on my daily carry MacBook Pro: because of a bug
in Emacs[1], the colors don't match what's specified in Ethan
Schoonover's original Solarized theme. There's thankfully a knob on
the theme to fix things, so on macOS turn it on.

(Based on the docs, `solarized-broken-srgb` should be 't by default on
macOS; weird that it isn't on my machine)

[1] https://debbugs.gnu.org/cgi/bugreport.cgi?bug=8402