Switch to using LSP
LSP is a little more flexible than using eclim, and actually supports
a lot more than just Java -- which means a lot of goodness when
editing stuff in other languages, including actual flycheck and symbol
lookups.

I've been using this at work, so it seems natural to apply it to my
non-work config.