More LSP quality-of-life improvements
Override the bindings for xref-find-{definitions,references} to use
the helper functions in lsp-ui directly, as I've found out that the
integration with xref tends to be weird/flakey on my set up
sometimes.

As well, for lsp-java enable on-type formatting and indentation,
hopefully overriding java-mode's config.

Finally, ensure company-lsp is loaded after lsp-mode.