Replace mood-line glyphs with my own custom set.
Specifically, I need a different glyph for :buffer-modified, since the
package-provided glyph for it isn't compatible with Berkeley
Mono. Specifically, the glyph has to be rendered using a different
font as Berkeley Mono doesn't have it defined. On my work machine, the
substitute font for the glyph (Andale Mono) happens to have the same
metrics (or similar enough ones) as Berkeley Mono, but on my personal
machine the substitute font that gets selected (Adelle Sans
Devanagari) is too "tall".

Instead of having to install another font, just use a different glyph
altogether that Berkeley Mono supports.