Properly configure system-packages-package-manager on macOS
The order of packages being loaded can get in a situation where
use-package-ensure-system-package is broken -- we don't explicitly set
the package manager (i.e. `system-packages-package-manager`), but also
we might not have brew's bin directory in exec-path.

So, explicitly set this up.

This is a fix for occasional start up issues I've been seeing where
some packages fail to load at start up (due to `demand: t`) with an
error like:

  alert/:config: install not supported in nil

which makes no sense; however, this[1] issue in the use-package repo
points to that error message as being about
`use-package-ensure-system-package` and the configured
`system-packages-package-manager` being `nil`.

---
[1] https://github.com/jwiegley/use-package/issues/698