Match ANSI colors with theme.
The anti-zenburn theme doesn't actually properly set `ansi-colors`;
the way it does it is through `ansi-color-names-vector`, which is
obsolete and so ignored -- also it maps colors in the reverse order of
what it's supposed to be. This means that we end up getting the wrong
color, making things unreadable (e.g. a bright green against a grey
background).

Fix it up instaed so we get appropriate color tints.