Replace bespoke JDT advice with one from eglot-java
I'm not going to use eglot-java fully, only bits and pieces. I don't
want eglot-java's auto-installation of JDTLS nor do I want it messing
with my eglot-server-programs config. However, I do want to use the
patch that it already has (and in fact where the work wiki sourced it
from), since that's more likely to be updated in the future.