Make vertico-posframe work with the latest vertico.
The latest vertico refactors things by renaming
`vertico--resize-window` to `vertico--resize` -- vertico-posframe
depends on that by providing a CLOS method override.

The symptom here is that in the latest vertico, instead of keeping the
minibuffer 1 line tall we get the default vertico behavior of resizing
the minibuffer -- for vertico-posframe this is obviously not what we
want.