Remove wonky/unnecessary override.
Was initially going to use this to put in a specific icon in the
notification message, but I think that's not doable overall on
macOS (either with terminal-notifier or via AppleScript). Plus, this
is the wrong way to override -- should've used an advice.

Since this does nothing different from the real implementation of
`gnus-desktop-notify-alert`, remove it altogether.