DWIM: If viewing a GitHub PR notification, jump to Pull Request.
When visiting a notification about a GitHub pull request,
automatagically jump to the PR when hitting C-<tab> from the summary;
otherwise, use the existing behavior of jumping to the first link in
the article.

This makes it easier for me to jump to the full PR if I want to review
it etc.