Ignore anything outside of ~/projects
With the exception of ~/.emacs.init/ ignore anything that is outside
of the ~/projects directory. This means that when I jump into a
definition into e.g. a Java source file decompiled from .class files,
or a source external Rust crate, Projectile doesn't deem it as a
project even if there are project marker files (Cargo.toml for the
external Rust crate, for instance), removing noise in my known
projects list.