L&F: Go back to modus-vivendi-tinted
Also going back to a generally dark system appearance. Generally,
anti-zenburn was good, but its low contrast started to annoy me, and
I've been searching for a less low contrast alternative similar to the
XEmacs theme -- which exists, though it works badly with mood-line, so
I opted not to use that.

Go back to using modus-vivendi-tinted, since it works well for me.