Fix keybinding breakage
I broke my prefix map in d3abe13a, as I forgot to `(provide
'jmi-keybindings)` which in turn suppressed loading a few
packages (including elfeed).

As I actually created that pseudo feature in 47472e6d as a way to
ensure that the prefix map exists before other use-package
declarations depend on it, I thought it best to just bite the bullet
and move the definition of the prefix map to main.el -- ensuring it
will always be loaded first anyway, so all other use-package
declarations can expect it to be there. This has the added bonus of
not needing the pseudo feature `jmi-keybindings`, cleaning things up a
bit.