Defer as much as possible to speed up startup
Previously, `M-x emacs-init-time` reported about ~4 seconds to
startup, as I'm unnecessarily loading a lot of things
immediately (such as the various language modes). Defer things as much
as possible so that packages instead get loaded when used instead of
at startup.

This in turn speeds up `M-x emacs-init-time` to ~1.5 seconds, which
still leaves a bit to be improved -- but it's a good start, and a
pretty good improvement.

NB: Actual wall clock time is closer to 5 seconds after speed
up (closer to 10 seconds for the base case), and a chunk of that seems
to be because of the Org Roam DB sync. Will investigate that later.