Move setting up org-roam autosync to end of startup screen
org-roam-db-autosync-mode sets up a bunch of hooks which end up
getting triggered by the startup screen. Particularly, it hooks
find-file, which org-agenda triggers since the underlying org files
are opened to scan for TODO items. So defer setting it up to after my
startup screen has run through.

This creates a minor issue in that the generated agenda may
potentially be incomplete, potentially missing a few new org files,
but that should be fine as it's unlikely I'll manually add files
while Emacs is not running.

Overall, this should then improve startup times.