Increase kill ring size
Should've done this way way earlier. It's not much of an issue for me
-- I don't mind if a few bits wind up off the end of the kill ring
necessitating me re-killing a bit of text -- but it's nice to have a
large history that I can rely on. I've been using Alfred's clipboard
history this way, so it's nice to have the same/similar parity in
terms of size inside of Emacs.