Move mail sync binding definition back to mbsync use-package clause
In my forks of my config for work, I have this set up in keybindings
since I also use a DWIM method that does a workspace sync (i.e.
hitting `<my prefix> f f` triggers either a mailbox sync or a
workspace sync, depending on which buffer I'm in). However, I don't
need the DWIM method at the moment, and when byte-compiling my config
and having a bunch of things deferred, the keybinding ends up getting
created with a dangling function.

Fix that so that it works by returning the binding definition back to
the mbsync use-package clause where it used to sit.