Hack around macOS Private Wi-Fi hostname changes
When Private Wi-Fi is enabled on macOS, either with rotating or fixed
MAC addresses, the OS obfuscates the hostname sent in DHCP option 12.
This in turn potentially affects what hostname the machine sees, and
is also different behavior from when connected over Ethernet.

This means that when a file is currently modified while connected to
Wi-Fi the lock symlink that Emacs creates will point to a different
system-name than when subsequently connected to Ethernet, even for the
same Emacs session; this in turn pops up a spurious prompt that the
file is being edited by someone else (which it is not).

The differences to hostnames in DHCP option 12 being sent are as
follows:

   - When Private Wi-Fi is set to "Rotating", option 12 is not sent/no
     host name is sent (not specifically sure which, but the behavior
     is equivalent). This then makes the machine fall back to using a
     .local address (e.g. mymachine.local) versus when connected over
     Ethernet where potentially the machine gets a FQDN (e.g.
     mymachine.somedomain.foo)

   - When Private Wi-Fi is set to "Fixed", option 12 is set to "Mac"
     or a similar non-descript hostname.

Handle this by advicing ask-user-about-lock to DWIM at least when
Private Wi-Fi is set to Rotating by ignoring the domain suffix and
treating this as though the user said to 's'teal the lock. This of
course is a little lax, and it will false-positive in some cases (e.g.
on a remote filesystem) -- I've minimized the chances of false
positives somewhat by also checking the PID of the other user and
comparing to our session.