Fix grokking PID for other-user
In 6a819f2e, when we check that the other-user on the lock has the
same PID as us, I've assumed that other-user is of a similar format as
the lock symlink value; e.g.

    user@host.789:123456

where 123456 is the boot time, and 789 is the PID -- this is the
format in the File Locks section of the Emacs manual.

However, what is passed in ask-user-about-lock is a human-readable
version, which looks more like

    user@host (pid 789)

Additionally, `last` returns the last element of the list _as a list_,
which ended up erroring during the string-equal comparisons.

So, fix those issues and make it actually work.