Add marginalia
Pretty useful, especially when looking at buffers -- marginalia
adds the (elided) path and major mode in buffer windows.

As specified by the package README though, the mode should be enabled
at package :init time -- it puzzles me why it needs to be this early,
but doing it in :config yields weirdness where the minor mode is
enabled but not actually working.